Add files via upload
Original upload after I got the prototype working. this is just the main code that loops forever reading sensors and printing and storing them to the Maria DB. This code has to be run on the RPi.  Other files were updated too to get things working but will need to be grabbed off the RPi to upload here in this repository.  The whole code set was from the original pull from the RPi weather station project and then I put in my own mods.